refactor: Rename handlers and move to controllers for better organization
